Improvements to GEQDSK reading of SCENE output
The poloidal flux is only defined up to a constant offset,
so the convention of setting flux to zero on the magnetic axis
in SCENE led to problems when fitting coil currents.
The 2D psi constraint now ignores constant offset differences.

Several places assumed that X-points could be found, but now
continue without.

The toroidal current can be calculated from the input file,
but was not previously used in the first solve. The toroidal
current is now passed into solve, skipping the need for critical point
finding.