Add Equilibrium.Btor function
Surprisingly complicated. Needs to distinguish between inside
and outside the separatrix, so needs a core mask function
which can be evaluated at any point.

Added code to _updatePlasmaPsi which finds O,X-points
and calculates the masking function. This is stored as an array
self.mask, and a spline self.mask_func.

This results in some removal of duplicate code, but means that
_updatePlasmaPsi must be called. For example if coil currents
are changed, but the plasma is not re-solved, then the result
will not be quite right.