Fix calculation of Equilibrium.psi_bndry when no X-points
For fixed boundary calculations without X-points, the psi_bndry
value was being set to None. This resulted in calculations which
depend on this value, like poloidal beta, failing.

Using the boundary value is only valid if a fixed boundary condition
is being used, so if free boundary then psi_bndry is still None.